browserstack example (#1409)
* beginning of getting browserstack working

* More configs for browserstack

* runner has been moved

* creating test suite helper

* fixes and moving some stuff around

* More fixes. Allow the user to pick and choose what test suites to run.

This will make it easier for the user if they are switching test
runners frequently. I didn’t like adding all the tests in one array per
configuration. This fixes that.

Also allows the user to pick and choose what test suites apply to them.

* lint fix

* Adding reporter for tests, changed admin email, adding npm scripts for running tests

* Allow Allure reporting to be optional

* Primer for switching login/checkout tests to register/checkout

* Login tests, are now register tests

This also removes the dependency to having to create a user prior to
running

* Adding reporting to browserstack conf

Also, added npm script for opening report. Another thing I noticed,
browserstack tunnel sometimes hangs. So added process killer.

* Updating test descriptions

* lint fix

* updating npm package

* Addresses login test with hardcoded user assertion

* This allows the user to use dynamic ids in tests

* updated login test to pass local

* updated test packages

* updated from development

- use REACTION_USER/AUTH if available.

* updated versions

* updated user data
