+++ .meteor/versions
@@ -1,4 +1,4 @@
-accounts-base@1.2.11
+accounts-base@1.2.12_1
 accounts-facebook@1.0.10
 accounts-google@1.0.10
 accounts-oauth@1.1.13
@@ -53,20 +53,20 @@ cfs:upload-http@0.0.20
 cfs:worker@0.1.4
 check@1.2.3
 chuangbo:cookie@1.1.0
-coffeescript@1.2.4_1
+coffeescript@1.2.4_2
 dburles:factory@1.1.0
 ddp@1.2.5
-ddp-client@1.3.1
+ddp-client@1.3.1_1
 ddp-common@1.2.6
 ddp-rate-limiter@1.0.5
-ddp-server@1.3.10
+ddp-server@1.3.10_1
 deps@1.0.12
 diff-sequence@1.0.6
 dispatch:mocha@0.0.9
 ecmascript@0.5.8_1
 ecmascript-runtime@0.3.14_1
 ejson@1.0.12
-email@1.1.17
+email@1.1.17_1
 es5-shim@4.6.14_1
 facebook@1.2.9
 fastclick@1.0.12
@@ -105,7 +105,7 @@ meteorhacks:meteorx@1.4.1
 meteorhacks:picker@1.0.3
 meteorhacks:ssr@2.2.0
 meteorhacks:subs-manager@1.6.4
-minifier-css@1.2.14
+minifier-css@1.2.14_1
 minifier-js@1.2.14_1
 minimongo@1.0.17
 mobile-experience@1.0.4
@@ -113,17 +113,17 @@ mobile-status-bar@1.0.12
 modules@0.7.6_1
 modules-runtime@0.7.6_1
 momentjs:moment@2.15.1
-mongo@1.1.12_1
+mongo@1.1.12_5
 mongo-id@1.0.5
 mongo-livedata@1.0.12
 mrt:later@1.6.1
-npm-bcrypt@0.9.1
-npm-mongo@1.5.50
+npm-bcrypt@0.9.1_1
+npm-mongo@2.2.10_1
 oauth@1.1.11
 oauth-encryption@1.2.0
 oauth1@1.1.10
 oauth2@1.1.10
-observe-sequence@1.0.12
+observe-sequence@1.0.13
 ongoworks:security@2.0.1
 ordered-dict@1.0.8
 practicalmeteor:chai@2.1.0_1
@@ -154,7 +154,7 @@ templating-compiler@1.2.15
 templating-runtime@1.2.15
 templating-tools@1.0.5
 tmeasday:check-npm-versions@0.3.1
-tmeasday:publish-counts@0.7.3
+tmeasday:publish-counts@0.8.0
 tracker@1.1.0
 twitter@1.1.12
 ui@1.0.12
