+++ tests/acceptance-tests/elements/element-ids.yml
@@ -1,0 +1,3 @@
+login_dropdown: Sign In
+login_email_fld_id: email
+login_pw_fld_id: password
