+++ tests/acceptance-tests/test/specs/payment-processors/authnet/authorizenet-void.app-test.js
@@ -1,9 +1,10 @@
 "use strict";
 const yaml = require("js-yaml");
 const fs   = require("fs");
-const shopUser = require("../../lib/user-shop-actions.js");
-const userDo = require("../../lib/basic-user-actions.js");
-const adminUser = require("../../lib/admin-order-actions.js");
+const expect = require("chai").expect;
+const shopUser = require("../../../../lib/user-shop-actions.js");
+const userDo = require("../../../../lib/basic-user-actions.js");
+const adminUser = require("../../../../lib/admin-order-actions.js");
 
 
 beforeEach(function () {
@@ -13,9 +14,9 @@ beforeEach(function () {
 });
 
 
-describe("example payment refund test", function () {
+describe("authorize net void test", function () {
   const eleMap = yaml.safeLoad(fs.readFileSync("./tests/acceptance-tests/elements/element-map.yml", "utf8"));
-  it("verify user can refund with example payment", function () {
+  it("verify user can void an order with authorize net", function () {
     userDo.UserActions.userLogin("admin");
     browser.pause("5000");
     userDo.UserActions.refreshShop();
@@ -29,22 +30,24 @@ describe("example payment refund test", function () {
     shopUser.checkForAddress();
     // free shipping option
     browser.click(eleMap.free_shipping);
-    browser.waitForEnabled(eleMap.example_payment, 5000);
-    browser.click(eleMap.example_payment);
-    shopUser.examplePaymentInfo();
-    browser.pause("5000");
-    browser.click(eleMap.example_payment_complete_order_btn);
+    browser.waitForEnabled(eleMap.authorizenet, 5000);
+    browser.click(eleMap.authorizenet);
+    shopUser.authorizeNetPaymentInfo();
+    browser.click(eleMap.authorizenet_complete_order_btn);
     browser.pause("6000");
+    // navigate to orders page
     browser.click(eleMap.orders_page_btn);
     browser.pause("5000");
     browser.click(eleMap.first_order_new_btn);
-    browser.waitForEnabled(eleMap.approve_btn, 5000);
+    adminUser.voidAmount();
     browser.click(eleMap.approve_btn);
-    browser.waitForEnabled(eleMap.capture_payment_btn, 5000);
+    browser.waitForVisible("h2", 5000);
+    expect(browser.getText("h2")).to.equal("Applying a 100% discount will void / cancel this order with your payment provider");
+    browser.click(eleMap.apply_discount_btn);
+    browser.waitForVisible(eleMap.order_total_post_discount, 5000);
+    const getTotal = browser.getText(eleMap.order_total_post_discount);
+    expect(getTotal[0]).to.equal("$0.00");
     browser.click(eleMap.capture_payment_btn);
-    adminUser.refundAmount();
-    browser.click(eleMap.apply_refund_btn);
-    browser.waitForEnabled(eleMap.apply_refund_pop_up_btn, 5000);
-    browser.click(eleMap.apply_refund_pop_up_btn);
+    browser.pause("3000");
   });
 });
