+++ tests/acceptance-tests/test/specs/payment-processors/authnet/guest-authorizenet-checkout.app-test.js
@@ -2,8 +2,8 @@
 const yaml = require("js-yaml");
 const fs   = require("fs");
 const expect = require("chai").expect;
-const shopUser = require("../../lib/user-shop-actions.js");
-const userDo = require("../../lib/basic-user-actions.js");
+const shopUser = require("../../../../lib/user-shop-actions.js");
+const userDo = require("../../../../lib/basic-user-actions.js");
 
 beforeEach(function () {
   const browserConfig = yaml.safeLoad(fs.readFileSync("./tests/acceptance-tests/config/settings.yml", "utf8"));
@@ -12,9 +12,9 @@ beforeEach(function () {
 });
 
 
-describe("stripe guest checkout test", function () {
+describe("authorize net guest checkout test", function () {
   const eleMap = yaml.safeLoad(fs.readFileSync("./tests/acceptance-tests/elements/element-map.yml", "utf8"));
-  it("verify guest can checkout with stripe", function () {
+  it("verify guest can checkout with authorize net", function () {
     browser.pause("5000");
     userDo.UserActions.refreshShop();
     browser.click(eleMap.product);
@@ -22,18 +22,18 @@ describe("stripe guest checkout test", function () {
     browser.click(eleMap.red_option);
     browser.waitForEnabled(eleMap.add_to_cart, 5000);
     browser.click(eleMap.add_to_cart);
-    browser.waitForEnabled(eleMap.checkout_btn, 2000);
+    browser.waitForEnabled(eleMap.checkout_btn, 3000);
     browser.click(eleMap.checkout_btn);
     browser.waitForEnabled(eleMap.continue_as_guest, 5000);
     browser.click(eleMap.continue_as_guest);
     shopUser.userAddress();
     // free shipping option
     browser.click(eleMap.free_shipping);
-    browser.waitForEnabled(eleMap.stripe, 3000);
-    browser.click(eleMap.stripe);
-    shopUser.stripePaymentInfo();
-    browser.click(eleMap.stripe_complete_order_btn);
-    browser.waitForVisible("#order-status", 20000);
+    browser.waitForEnabled(eleMap.authorizenet, 5000);
+    browser.click(eleMap.authorizenet);
+    shopUser.authorizeNetPaymentInfo();
+    browser.click(eleMap.authorizenet_complete_order_btn);
+    browser.waitForVisible("#order-status", 10000);
     expect(browser.getText("#order-status")).to.equal("Your order is now submitted.");
   });
 });
