+++ tests/acceptance-tests/test/specs/payment-processors/braintree/braintree-refund.app-test.js
@@ -1,9 +1,10 @@
 "use strict";
 const yaml = require("js-yaml");
 const fs   = require("fs");
-const shopUser = require("../../lib/user-shop-actions.js");
-const userDo = require("../../lib/basic-user-actions.js");
-const adminUser = require("../../lib/admin-order-actions.js");
+const expect = require("chai").expect;
+const shopUser = require("../../../../lib/user-shop-actions.js");
+const userDo = require("../../../../lib/basic-user-actions.js");
+const adminUser = require("../../../../lib/admin-order-actions.js");
 
 
 beforeEach(function () {
@@ -13,9 +14,9 @@ beforeEach(function () {
 });
 
 
-describe("stripe refund test", function () {
+describe("braintree refund test", function () {
   const eleMap = yaml.safeLoad(fs.readFileSync("./tests/acceptance-tests/elements/element-map.yml", "utf8"));
-  it("verify user can refund with stripe", function () {
+  it("verify user can refund with braintree", function () {
     userDo.UserActions.userLogin("admin");
     browser.pause("5000");
     userDo.UserActions.refreshShop();
@@ -29,23 +30,27 @@ describe("stripe refund test", function () {
     shopUser.checkForAddress();
     // free shipping option
     browser.click(eleMap.free_shipping);
-    browser.waitForEnabled(eleMap.stripe, 5000);
-    browser.click(eleMap.stripe);
-    shopUser.stripePaymentInfo();
-    browser.waitForEnabled(eleMap.stripe_complete_order_btn, 5000);
-    browser.click(eleMap.stripe_complete_order_btn);
-    browser.pause("9000");
+    browser.waitForEnabled(eleMap.braintree, 5000);
+    browser.click(eleMap.braintree);
+    shopUser.braintreePaymentInfo();
+    browser.waitForEnabled(eleMap.braintree_complete_order_btn, 5000);
+    browser.click(eleMap.braintree_complete_order_btn);
+    browser.pause(5000);
     browser.click(eleMap.orders_page_btn);
-    browser.pause("5000");
+    browser.pause("15000");
     browser.click(eleMap.first_order_new_btn);
-    browser.waitForEnabled(eleMap.approve_btn, 5000);
+    browser.waitForEnabled(eleMap.approve_btn, 10000);
     browser.click(eleMap.approve_btn);
     browser.waitForEnabled(eleMap.capture_payment_btn, 5000);
     browser.click(eleMap.capture_payment_btn);
     adminUser.refundAmount();
     browser.click(eleMap.apply_refund_btn);
-    browser.pause("5000");
+    browser.pause(5000);
     browser.click(eleMap.apply_refund_pop_up_btn);
-    browser.pause("2000");
+    browser.pause(5000);
+    expect(browser.getText("h2"))
+    .to.equal("Braintree does not allow refunds until transactions are settled. This can take up to 24 hours. Please try again later.");
+    // error pop up ok btn
+    browser.click("button.swal2-confirm.styled");
   });
 });
