+++ tests/acceptance-tests/test/specs/payment-processors/braintree/register-braintree-checkout.app-test.js
@@ -2,8 +2,8 @@
 const yaml = require("js-yaml");
 const fs   = require("fs");
 const expect = require("chai").expect;
-const shopUser = require("../../lib/user-shop-actions.js");
-const userDo = require("../../lib/basic-user-actions.js");
+const shopUser = require("../../../../lib/user-shop-actions.js");
+const userDo = require("../../../../lib/basic-user-actions.js");
 
 
 beforeEach(function () {
@@ -13,10 +13,10 @@ beforeEach(function () {
 });
 
 
-describe("authorize net logged in checkout test", function () {
+describe("braintree register user and checkout test", function () {
   const eleMap = yaml.safeLoad(fs.readFileSync("./tests/acceptance-tests/elements/element-map.yml", "utf8"));
-  it("verify logged in user can checkout with authorize net", function () {
-    userDo.UserActions.userLogin("guest");
+  it("verify user can register and checkout with braintree", function () {
+    userDo.UserActions.registerUser();
     browser.pause("5000");
     userDo.UserActions.refreshShop();
     browser.click(eleMap.product);
@@ -24,15 +24,16 @@ describe("authorize net logged in checkout test", function () {
     browser.click(eleMap.red_option);
     browser.waitForEnabled(eleMap.add_to_cart, 5000);
     browser.click(eleMap.add_to_cart);
-    browser.waitForEnabled(eleMap.checkout_btn, 3000);
+    browser.waitForEnabled(eleMap.checkout_btn, 5000);
     browser.click(eleMap.checkout_btn);
     shopUser.checkForAddress();
     // free shipping option
     browser.click(eleMap.free_shipping);
-    browser.waitForEnabled(eleMap.authorizenet, 5000);
-    browser.click(eleMap.authorizenet);
-    shopUser.authorizeNetPaymentInfo();
-    browser.click(eleMap.authorizenet_complete_order_btn);
+    browser.waitForEnabled(eleMap.braintree, 5000);
+    browser.click(eleMap.braintree);
+    shopUser.braintreePaymentInfo();
+    browser.waitForEnabled(eleMap.braintree_complete_order_btn, 5000);
+    browser.click(eleMap.braintree_complete_order_btn);
     browser.waitForVisible("#order-status", 20000);
     expect(browser.getText("#order-status")).to.equal("Your order is now submitted.");
   });
