+++ tests/acceptance-tests/test/specs/payment-processors/example/guest-example-payment-checkout.app-test.js
@@ -2,9 +2,8 @@
 const yaml = require("js-yaml");
 const fs   = require("fs");
 const expect = require("chai").expect;
-const shopUser = require("../../lib/user-shop-actions.js");
-const userDo = require("../../lib/basic-user-actions.js");
-
+const shopUser = require("../../../../lib/user-shop-actions.js");
+const userDo = require("../../../../lib/basic-user-actions.js");
 
 beforeEach(function () {
   const browserConfig = yaml.safeLoad(fs.readFileSync("./tests/acceptance-tests/config/settings.yml", "utf8"));
@@ -13,10 +12,9 @@ beforeEach(function () {
 });
 
 
-describe("paypal logged in checkout test", function () {
+describe("example payment guest checkout test", function () {
   const eleMap = yaml.safeLoad(fs.readFileSync("./tests/acceptance-tests/elements/element-map.yml", "utf8"));
-  it("verify logged in user can checkout with paypal", function () {
-    userDo.UserActions.userLogin("guest");
+  it("verify guest can checkout with example payment", function () {
     browser.pause("5000");
     userDo.UserActions.refreshShop();
     browser.click(eleMap.product);
@@ -24,18 +22,19 @@ describe("paypal logged in checkout test", function () {
     browser.click(eleMap.red_option);
     browser.waitForEnabled(eleMap.add_to_cart, 5000);
     browser.click(eleMap.add_to_cart);
-    browser.waitForEnabled(eleMap.checkout_btn, 5000);
+    browser.waitForEnabled(eleMap.checkout_btn, 2000);
     browser.click(eleMap.checkout_btn);
-    shopUser.checkForAddress();
+    browser.waitForEnabled(eleMap.continue_as_guest, 5000);
+    browser.click(eleMap.continue_as_guest);
+    shopUser.userAddress();
     // free shipping option
     browser.click(eleMap.free_shipping);
-    browser.waitForEnabled(eleMap.paypal, 5000);
-    browser.click(eleMap.paypal);
-    shopUser.paypalPaymentInfo();
-    browser.waitForEnabled(eleMap.paypal_complete_order_btn, 5000);
-    browser.click(eleMap.paypal_complete_order_btn);
-    // paypal takes forever 45sec wait
-    browser.waitForVisible("#order-status", 45000);
+    browser.waitForEnabled(eleMap.example_payment, 5000);
+    browser.click(eleMap.example_payment);
+    shopUser.examplePaymentInfo();
+    browser.waitForEnabled(eleMap.example_payment_complete_order_btn, 5000);
+    browser.click(eleMap.example_payment_complete_order_btn);
+    browser.waitForVisible("#order-status", 10000);
     expect(browser.getText("#order-status")).to.equal("Your order is now submitted.");
   });
 });
