+++ tests/acceptance-tests/test/specs/permissions/dashboard-permissions.app-test.js
@@ -2,7 +2,8 @@
 const yaml = require("js-yaml");
 const fs   = require("fs");
 const expect = require("chai").expect;
-const adminUser = require("../../lib/basic-user-actions.js");
+const adminUser = require("../../../lib/basic-user-actions.js");
+const resetUser = require("../../../lib/user-permissions-reset.js");
 
 beforeEach(function () {
   const browserConfig = yaml.safeLoad(fs.readFileSync("./tests/acceptance-tests/config/settings.yml", "utf8"));
@@ -11,28 +12,31 @@ beforeEach(function () {
 });
 
 afterEach(function () {
-
+  resetUser.resetUserPermissions();
 });
 
-describe("core permissions test", function () {
+describe("dashboard permissions test", function () {
   const eleMap = yaml.safeLoad(fs.readFileSync("./tests/acceptance-tests/elements/element-map.yml", "utf8"));
-  it("verify admin adding dashboard permissions to guest", function () {
+  it("verify admin adding dashboard permissions to guest user", function () {
     adminUser.UserActions.userLogin("admin");
     // navigate to dashboard
     browser.click(eleMap.dashboard_btn);
     browser.pause("5000");
+    // select guest user from user page
     browser.click(eleMap.db_accounts);
     browser.pause("5000");
     browser.click(eleMap.account_manage_btn);
     browser.pause("5000");
     browser.click(eleMap.perm_dashboard_chkb);
-    browser.click(eleMap.perm_shop_settings_chkb);
+    browser.pause("5000");
     adminUser.UserActions.userLogout();
+    browser.pause("5000");
+    // login as guest user
     adminUser.UserActions.userLogin("guest");
+    // user should be able to enter the dashboard now
     browser.click(eleMap.dashboard_btn);
     browser.pause("5000");
-    // verify Core settings option/button is present
-    expect(browser.getText(eleMap.core_settings_btn_txt)).to.equal("Core");
+    expect(browser.getText(eleMap.dashboard_hdr_txt)).to.equal("Dashboard");
     adminUser.UserActions.userLogout();
   });
 });
