+++ tests/acceptance-tests/test/specs/permissions/shop-settings-permissions.app-test.js
@@ -2,8 +2,7 @@
 const yaml = require("js-yaml");
 const fs   = require("fs");
 const expect = require("chai").expect;
-const adminUser = require("../../lib/basic-user-actions.js");
-const resetUser = require("../../lib/user-permissions-reset.js");
+const adminUser = require("../../../lib/basic-user-actions.js");
 
 beforeEach(function () {
   const browserConfig = yaml.safeLoad(fs.readFileSync("./tests/acceptance-tests/config/settings.yml", "utf8"));
@@ -12,31 +11,28 @@ beforeEach(function () {
 });
 
 afterEach(function () {
-  resetUser.resetUserPermissions();
+
 });
 
-describe("dashboard permissions test", function () {
+describe("core permissions test", function () {
   const eleMap = yaml.safeLoad(fs.readFileSync("./tests/acceptance-tests/elements/element-map.yml", "utf8"));
-  it("verify admin adding dashboard permissions to guest user", function () {
+  it("verify admin adding dashboard permissions to guest", function () {
     adminUser.UserActions.userLogin("admin");
     // navigate to dashboard
     browser.click(eleMap.dashboard_btn);
     browser.pause("5000");
-    // select guest user from user page
     browser.click(eleMap.db_accounts);
     browser.pause("5000");
     browser.click(eleMap.account_manage_btn);
     browser.pause("5000");
     browser.click(eleMap.perm_dashboard_chkb);
-    browser.pause("5000");
+    browser.click(eleMap.perm_shop_settings_chkb);
     adminUser.UserActions.userLogout();
-    browser.pause("5000");
-    // login as guest user
     adminUser.UserActions.userLogin("guest");
-    // user should be able to enter the dashboard now
     browser.click(eleMap.dashboard_btn);
     browser.pause("5000");
-    expect(browser.getText(eleMap.dashboard_hdr_txt)).to.equal("Dashboard");
+    // verify Core settings option/button is present
+    expect(browser.getText(eleMap.core_settings_btn_txt)).to.equal("Core");
     adminUser.UserActions.userLogout();
   });
 });
