+++ tests/acceptance-tests/test/specs/smoke-tests/simple-login.app-test.js
@@ -2,23 +2,35 @@
 const yaml = require("js-yaml");
 const fs   = require("fs");
 const expect = require("chai").expect;
+const getId = require("../../../lib/get-elements.js");
 
 beforeEach(function () {
   const browserConfig = yaml.safeLoad(fs.readFileSync("./tests/acceptance-tests/config/settings.yml", "utf8"));
   const baseUrl = browserConfig.base_url.toString();
   browser.url(baseUrl);
+  // browser.getSession().then(function (sessionid) {
+  //   browser.sessionID = sessionid.id_;
+  // });
 });
 
 describe("simple login test", function () {
   it("verify user is able to login - and verifies user name in dropdown", function () {
     const eleMap = yaml.safeLoad(fs.readFileSync("./tests/acceptance-tests/elements/element-map.yml", "utf8"));
+    const eleIds = yaml.safeLoad(fs.readFileSync("./tests/acceptance-tests/elements/element-ids.yml", "utf8"));
     const usrData = yaml.safeLoad(fs.readFileSync("./tests/acceptance-tests/config/user-data.yml", "utf8"));
+
+    // default to process env if we've got that
+    const adminEmail = process.env.REACTION_EMAIL || usrData.admin_email;
+    const adminPassword = process.env.REACTION_AUTH || usrData.admin_pw;
+    const adminUserName = process.env.REACTION_USER || usrData.admin_user;
+
     browser.pause("5000");
     browser.click(eleMap.login_dropdown_btn);
-    browser.setValue(eleMap.login_email_fld, usrData.admin_email);
-    browser.setValue(eleMap.login_pw_fld, usrData.admin_pw);
+    browser.pause(5000);
+    browser.setValue(getId.retId(eleIds.login_email_fld_id), adminEmail);
+    browser.setValue(getId.retId(eleIds.login_pw_fld_id), adminPassword);
     browser.click(eleMap.login_btn);
     browser.pause("5000");
-    expect(browser.getText("#logged-in-display-name")).to.equal("Admin");
+    expect(browser.getText("#logged-in-display-name")).to.equal(adminUserName);
   });
 });
